+++ src/core/layers/InteractionCaptureLayer.tsx
@@ -57,7 +57,7 @@ export default class InteractionCaptureLayer extends React.Component<Props, Stat
   render() {
     return (
       <div
-        className='interaction-capture'
+        className='interaction-capture interaction-capture-layer'
         onMouseDown={this._onMouseDown}
         onMouseUp={this._onMouseUp}
         onMouseMove={this._onMouseMove}